Move CORS tests all together
