+++ cliquet/tests/resource/test_views.py
@@ -190,91 +190,6 @@ class InvalidBodyTest(FakeAuthentMixin, BaseWebTest):
         self.assertIn('escape sequence', resp.json['message'])
 
 
-class CORSHeadersTest(FakeAuthentMixin, BaseWebTest):
-    def setUp(self):
-        super(CORSHeadersTest, self).setUp()
-        self.headers['Origin'] = 'notmyidea.org'
-
-        response = self.app.post_json(self.collection_url,
-                                      MINIMALIST_RECORD,
-                                      headers=self.headers,
-                                      status=201)
-        self.record = response.json
-
-    def test_present_on_hello(self):
-        response = self.app.get('/',
-                                headers=self.headers,
-                                status=200)
-        self.assertIn('Access-Control-Allow-Origin', response.headers)
-
-    def test_present_on_single_record(self):
-        response = self.app.get(self.get_item_url(),
-                                headers=self.headers)
-        self.assertIn('Access-Control-Allow-Origin', response.headers)
-
-    def test_present_on_deletion(self):
-        response = self.app.delete(self.get_item_url(),
-                                   headers=self.headers)
-        self.assertIn('Access-Control-Allow-Origin', response.headers)
-
-    def test_present_on_unknown_record(self):
-        url = self.get_item_url('unknown')
-        response = self.app.get(url,
-                                headers=self.headers,
-                                status=404)
-        self.assertIn('Access-Control-Allow-Origin', response.headers)
-
-    def test_present_on_invalid_record_update(self):
-        body = {'name': 42}
-        response = self.app.patch_json(self.get_item_url(),
-                                       body,
-                                       headers=self.headers,
-                                       status=400)
-        self.assertIn('Access-Control-Allow-Origin', response.headers)
-
-    def test_present_on_successful_creation(self):
-        response = self.app.post_json(self.collection_url,
-                                      MINIMALIST_RECORD,
-                                      headers=self.headers,
-                                      status=201)
-        self.assertIn('Access-Control-Allow-Origin', response.headers)
-
-    def test_present_on_invalid_record_creation(self):
-        body = {'name': 42}
-        response = self.app.post_json(self.collection_url,
-                                      body,
-                                      headers=self.headers,
-                                      status=400)
-        self.assertIn('Access-Control-Allow-Origin', response.headers)
-
-    def test_present_on_readonly_update(self):
-        with mock.patch('cliquet.tests.testapp.views.'
-                        'MushroomSchema.is_readonly',
-                        return_value=True):
-            body = {'name': 'Amanite'}
-            response = self.app.patch_json(self.get_item_url(),
-                                           body,
-                                           headers=self.headers,
-                                           status=400)
-        self.assertEqual(response.json['message'], 'Cannot modify name')
-        self.assertIn('Access-Control-Allow-Origin', response.headers)
-
-    def test_present_on_unauthorized(self):
-        self.headers.pop('Authorization', None)
-        response = self.app.post_json(self.collection_url,
-                                      MINIMALIST_RECORD,
-                                      headers=self.headers,
-                                      status=401)
-        self.assertIn('Access-Control-Allow-Origin', response.headers)
-
-    def test_present_on_internal_error(self):
-        with mock.patch('cliquet.resource.BaseResource.get_records',
-                        side_effect=ValueError):
-            response = self.app.get('/mushrooms',
-                                    headers=self.headers, status=500)
-        self.assertIn('Access-Control-Allow-Origin', response.headers)
-
-
 class ConflictErrorsTest(FakeAuthentMixin, BaseWebTest):
     def setUp(self):
         super(ConflictErrorsTest, self).setUp()
